Create a new theme folder. Move dotfiles too.
The themes are now in a themes folder. There are a couple of sections in
the setup script that copies them over. The terminal one has to be
manually selected.

The dot files are moved to a dotfiles folder and the setup script has
been modified to link them there instead.